Address review feedback on the I2S port count and the LED pin check
- Define M5UNIFIED_I2S_PORT_COUNT as 2 on cores that predate soc_caps.h
  (Arduino core 1.x), which are always the classic ESP32. The macro used
  to stay undefined there, so "#if M5UNIFIED_I2S_PORT_COUNT >= 2"
  evaluated to 0 and the raw register path silently stayed on I2S0 while
  the driver used I2S_NUM_1 (e.g. HAT SPK2 and Module Display/RCA
  audio). This is a pre-existing compatibility defect rather than a
  regression of this branch: the previous "#if SOC_I2S_NUM >= 2"
  evaluated the same way on those cores.
- Emit #error only when the target declares I2S support
  (SOC_I2S_SUPPORTED) but the port count cannot be resolved, so that
  targets without an I2S peripheral can include the header freely.
- Simplify the LED strip data pin range check; the value is already
  known to be non-negative at that point.

+++ src/utility/led/LED_Strip_Class.cpp
@@ -131,7 +131,7 @@ namespace m5
       return true;
     }
 
-    if ((size_t)_config.pin_data >= GPIO_NUM_MAX || !GPIO_IS_VALID_OUTPUT_GPIO((gpio_num_t)_config.pin_data)) {
+    if (_config.pin_data >= GPIO_NUM_MAX || !GPIO_IS_VALID_OUTPUT_GPIO((gpio_num_t)_config.pin_data)) {
       return false;
     }
     // Normalize the data pin as push-pull output low before handing it to the RMT peripheral.
